Fix notification bug
